requirements.txt
llama-cpp-python[server]
fastapi
uvicorn
python-dotenv
